server.js tests
